Refactor main.py with sound effects and loading video
Refactor main.py to enhance functionality and add sound effects.

+++ main.py
@@ -1,25 +1,53 @@
-import tkinter
 import splashscreen_engine as splash
 import pygame
 import keyboard
+
 import threading
 import os
 import random
+import datetime
+import tkinter
 
 import DataHandling
 import AdditionalUi
 
-data = DataHandling.Uid(username="IDK")
+# Initialize Mixer
+pygame.mixer.init()
+
+# Sounds
+HorrorSound = pygame.mixer.Sound("Assets/Audios/Horror.mp3")
+SignalLost = pygame.mixer.Sound("Assets/Audios/SignalLost.mp3")
+ShutterSound = pygame.mixer.Sound("Assets/Audios/CameraShutter.mp3")
+PopUps = pygame.mixer.Sound("Assets/Audios/PopUps.mp3")
+winCrash = pygame.mixer.Sound("Assets/Audios/windowCrashed.mp3")
+
+# Music Channel
+channel = pygame.mixer.Channel(1)
+channel.set_volume(0.5)
+
+# Sound Effects
+channelEffects = pygame.mixer.Channel(2)
+channelEffects.set_volume(0.6)
+
+# Play Horror music
+channel.play(HorrorSound,loops=-1)
+
+# Data handling
+data = DataHandling.Uid()
 if not data.is_created():
-    data.make()
+    AdditionalUi.AskName() # Asks name and make Unique ID once only
 DataHandling.update_last_played()
 DataHandling.save_database()
 
+# Time Calculator
+start_time = datetime.datetime.now()
 
-
+# Setting up screen
 screen = splash.Screen()
 screen.size(fullscreen=True) # Add more fun
 screen.start()
+loadingVid = splash.BackgroundVideo(screen,"Assets/Videos/loadingScreen.mp4")
+loadingVid.play()
 
 # Default Font
 font = "Orbitron Bold"
@@ -43,14 +71,15 @@ init_messages = [
 ]
 
 # Disable Keyboard
-hook = keyboard.hook(lambda x: keyboard.block_key(x.name))
+if DataHandling.get_username() != "OWNER":
+    hook = keyboard.hook(lambda x: keyboard.block_key(x.name))
 
 # Animate Text
 for i in range(0,5):
     for k in range(0,20):
         value+=1
         bar.set_progress(value)
-        screen.wait(0.1)
+        screen.wait(0.3)
     text.edit(text=init_messages[i])
     print(value)
 # screen.wait(2)
@@ -59,10 +88,15 @@ for i in range(0,5):
 text.hide()
 bar.hide()
 
+# Stop Old Video
+loadingVid.delete()
+
 # # Add Videos after Init
 bg_video = splash.BackgroundVideo(screen,"Assets/Videos/1.mp4")
 bg_video.play()
 
+# Play Sound
+channelEffects.play(SignalLost)
 
 
 
@@ -72,22 +106,30 @@ while bg_video.is_playing:
         if event.type == pygame.KEYDOWN:
             pass
 
-# Show opening camera
-text.edit(text="Finding Camera...")
-text.show()
+channelEffects.fadeout(2)
 
-import Camera
+if DataHandling.get_username() != "OWNER":
+    # Show opening camera
+    text.edit(text="Finding Camera...")
+    text.show()
 
-camera = Camera.OpenCamera()
-if camera.success:
-    text.edit(text="Opening Camera...")
-    screen.wait(5)
-    threading.Thread(target=camera.show).start()
-    screen.wait(10)
-    camera.stop()
-else:
-    text.edit(text="No Camera Found")
-    screen.wait(5)
+    import Camera
+
+    camera = Camera.OpenCamera()
+    DataHandling.save_camera(camera.success)
+    if camera.success:
+        channelEffects.play(ShutterSound)
+        text.edit(text="Opening Camera...")
+        screen.wait(5)
+        threading.Thread(target=camera.show).start()
+        screen.wait(11)
+        camera.stop()
+    else:
+        text.edit(text="No Camera Found")
+        screen.wait(5)
+
+# Pop Up sound
+channelEffects.play(PopUps)
 
 # Show Lots of tkinter popups
 text.edit(text="Close Main Window To Exit")
@@ -107,19 +149,37 @@ main_window.mainloop()
 
 text.edit(text="THANKS FOR USING THIS FAKE VIRUES")
 AdditionalUi.RateUs()
+
+# Playtime Seconds
+exit_time = datetime.datetime.now()
+time_played = (exit_time - start_time).total_seconds()
+DataHandling.update_playtime(time_played)
+
+# Save Database
 DataHandling.save_database()
+
 text.hide()
+
+# Show windows crashed (Fake - Just an image)
+bg_video.delete()
+
+screen.screen.fill((255,255,255))
+
 bg_image = splash.BackgroundImage(screen,"Assets/EndImage.png")
+channelEffects.play(winCrash)
 bg_image.set()
+
 
 screen.wait(6)
 
 screen.stop()
 
-# Enable Keyboard
-keyboard.unhook(hook)
+if DataHandling.get_username() != "OWNER":
+    # Enable Keyboard
+    keyboard.unhook(hook)
 
-os.system("shutdown /r /t 0")
+    # Restart System
+    os.system("shutdown /r /t 0")
